refactor(rmdx): shuffle around some classes
- Add a small amount more docs
