+++ packages/rmdx/src/main/exceptions/restricted-syntax-exception.ts
