+++ packages/rmdx/src/main/index.ts
@@ -4,9 +4,9 @@
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
  */
-export { MalformedMdxException } from './errors/malformed-mdx-exception.js'
-export { ProcessingException } from './errors/processing-exception.js'
-export { RestrictedSyntaxException } from './errors/restricted-syntax-exception.js'
+export { MalformedMdxException } from './exceptions/malformed-mdx-exception.js'
+export { ProcessingException } from './exceptions/processing-exception.js'
+export { RestrictedSyntaxException } from './exceptions/restricted-syntax-exception.js'
 export { AllowedComponents, AstNode, Renderer } from './interfaces.js'
 export { peek } from './peek.js'
 export { process } from './process.js'
